+++ BackEnd/src/main/java/com/phofor/phocaforme/chat/service/ChatRoomService.java
@@ -8,12 +8,12 @@ import java.util.List;
 public interface ChatRoomService {
 
     // 방 있는지 체크해서 방 만들어주기
-    ChatRoomResponseDto getChatRoomByBoardIdAndVisiterId(Integer boardId, CustomOAuth2User customOAuth2User);
+    ChatRoomResponseDto getChatRoomByBoardIdAndVisiterId(Long boardId, CustomOAuth2User customOAuth2User);
 
     // 유저의 채팅 내역 모두 가져오기
     List<ChatRoomResponseDto> getAllByOwnerIdOrVisiterId(CustomOAuth2User customOAuth2User);
 
     // 채팅방 만들기
 //    @Transactional
-//    ChatRoomResponseDto getChatRoomByBoardIdAndVisiterId(ChatRoomRequestDto chatRoomRequestDto, Integer boardId, String userId);
+//    ChatRoomResponseDto getChatRoomByBoardIdAndVisiterId(ChatRoomRequestDto chatRoomRequestDto, Long boardId, String userId);
 }
